Database Maintenance, checking for claim payments with no claimprocs would cause an error sometimes. Backported to: 15.2.12.0 Reviewed by: Jason
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@12530 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Misc/DatabaseMaintenance.cs
@@ -1389,7 +1389,7 @@ namespace OpenDentBusiness {
 					for(int i=0;i<table.Rows.Count;i++) {
 						log+="   "+Lans.g("FormDatabaseMaintenance","Date")+": "+PIn.Date(table.Rows[i]["CheckDate"].ToString()).ToShortDateString();
 						log+=", "+Lans.g("FormDatabaseMaintenance","Type")+": "+PIn.String(table.Rows[i]["ItemName"].ToString());
-						log+=", "+Lans.g("FormDatabaseMaintenance","Amount")+": "+PIn.Long(table.Rows[i]["CheckAmt"].ToString()).ToString("c");
+						log+=", "+Lans.g("FormDatabaseMaintenance","Amount")+": "+PIn.Double(table.Rows[i]["CheckAmt"].ToString()).ToString("c");
 						//Partial will always be blank
 						log+=", "+Lans.g("FormDatabaseMaintenance","Carrier")+": "+PIn.String(table.Rows[i]["CarrierName"].ToString());
 						log+=", "+Lans.g("FormDatabaseMaintenance","Clinic")+": "+PIn.String(table.Rows[i]["Description"].ToString());
